OKR Template Added
The OKR template can be used as a zero cost way of starting out. Tab one has a nice summary diagram of the cascade from Vision to Objectives to OKRs, the second tab has a template for building up your quarterly OKRs. Create a new tab for each team / individual and work on them together.